requirements.txt
prettytable
pyyaml
